+++ app/root.tsx
@@ -7,14 +7,7 @@ import {
   Scripts,
   ScrollRestoration,
 } from "@remix-run/react";
-import { Navbar } from "./components/Navbar";
 import { getAccessToken } from "./cookies.server";
-
-import styles from "./tailwind.css";
-
-export const links: LinksFunction = () => [
-  { rel: "stylesheet", href: styles },
-];
 
 export const meta: MetaFunction = () => ({
   charset: "utf-8",
@@ -36,19 +29,10 @@ export default function App() {
         <Links />
       </head>
       <body>
-        <div className="h-screen flex flex-col">
-          <header>
-            <div className="bg-primary p-10 text-center text-3xl">Bonvinlab image banner</div>
-            <Navbar />
-          </header>
-          <div className="grow m-10">
-            <Outlet />
-          </div>
-          <ScrollRestoration />
-          <Scripts />
-          <LiveReload />
-          <footer className="bg-primary text-center p-1">Proudly sponsored by BonVinLab funders</footer>
-        </div>
+        <Outlet />
+        <ScrollRestoration />
+        <Scripts />
+        <LiveReload />
       </body>
     </html>
   );
