+++ app/routes/_layout.builder.tsx
@@ -1,0 +1,5 @@
+export default function Builder() {
+    return (
+        <iframe src="/applications/haddock3" className="w-full h-full"></iframe>
+    )
+}
