+++ app/routes/_layout.jobs/$id/stdout.tsx
