+++ remix.config.js
@@ -7,5 +7,6 @@ module.exports = {
   // publicPath: "/build/",
   future: {
     unstable_tailwind: true,
+    v2_routeConvention: true,
   },
 };
